Move type to file
